Download shared API keys for Giphy and Tenor
Asking users to sign up for their own developer accounts to get API keys is annoying and causes friction. This update downloads a config.json file from a known location that contains its own set of API keys.

If a user does want to use their own set of keys (maybe the shared keys got blocked or rate limited), then they can provide those in the Command Preferences, and they'll be used instead.

The shared config keys are stored on a user's machine and re-used across extension launches to increase performance. On launch, we check the last modified time of the file and if it's older than a week, we download it fresh.